issue116 Bug fixes and fix tests

+++ Museum-Backend/src/main/java/ca/mcgill/ecse321/museum/service/ArtworkService.java
@@ -260,7 +260,7 @@ public class ArtworkService {
 
     // Delete loan if artwork is on loan
     List<Loan> loan = loanRepository.findLoanByArtwork(artwork);
-    if (loan != null || loan.size() > 0) {
+    if (loan != null && loan.size() > 0) {
       loanRepository.deleteLoanByArtwork(artwork);
     }
